+++ src/app/admin/(with-sidebar)/application-edit/_components/calendar/style/application-form-datepicker.css
