+++ src/app/admin/(with-sidebar)/application-edit/_components/recruitment/AdminRecruitmentInformation.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import {RecruitmentInfoEditRow} from '@/app/admin/(with-sidebar)/application-form/_components/recruitment/RecruitmentInfoEditRow';
-import {RecruitmentInfoViewRow} from '@/app/admin/(with-sidebar)/application-form/_components/recruitment/RecruitmentInfoViewRow';
-import {scheduleSections} from '@/constants/admin/admin-application-form';
+import {RecruitmentInfoEditRow} from '@/app/admin/(with-sidebar)/application-edit/_components/recruitment/RecruitmentInfoEditRow';
+import {RecruitmentInfoViewRow} from '@/app/admin/(with-sidebar)/application-edit/_components/recruitment/RecruitmentInfoViewRow';
+import {scheduleSections} from '@/constants/admin/admin-application-questions';
 import {RecruitmentInformation} from '@/schemas/admin/admin-recruitment-information.schema';
 
 import {formatRecruitmentDate} from '@/utils/formatDate';
