+++ src/app/admin/(with-sidebar)/application-edit/_components/recruitment/RecruitmentInfoViewRow.tsx
